make general error decorator, need to change pythonbackend
python backend captures errors, and then sends them back as a
dictionary. This is problematic because it leads to excessive try/except
clauses inside each process function, in order to stop errors from
hitting the backend. Relaxing it to be less aggressive in error
capturing should clean up the code.